rename crop_viewer to media_viewer
